requirements.txt
alpaca-py==0.43.2
annotated-types==0.7.0
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
msgpack==1.1.2
numpy==2.4.0
pandas==2.3.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sseclient-py==1.8.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
websockets==15.0.1
